cmd: refine the commoand-line usage and fix tiup run playground
