+++ mirror/fetch.sh
@@ -2,25 +2,24 @@
 
 for x in tidb tikv pd
 do
-	for i in {0..10}
+	for v in $@
 	do
-		wget https://download.pingcap.org/$x-v3.0.$i-darwin-amd64.tar.gz
-		tar -xzf $x-v3.0.$i-darwin-amd64.tar.gz
-		tar -C $x-v3.0.$i-darwin-amd64 -czf $x-v3.0.$i-darwin-amd64.tar.gz bin/$x-server
-		rm -rf $x-v3.0.$i-darwin-amd64
-		sha1sum $x-v3.0.$i-darwin-amd64.tar.gz | awk '{print $1}' > $x-v3.0.$i-darwin-amd64.sha1
+		wget https://download.pingcap.org/$x-$v-darwin-amd64.tar.gz
+		tar -xzf $x-$v-darwin-amd64.tar.gz
+		tar -C $x-$v-darwin-amd64 -czf $x-$v-darwin-amd64.tar.gz bin/$x-server
+		rm -rf $x-$v-darwin-amd64
+		sha1sum $x-$v-darwin-amd64.tar.gz | awk '{print $1}' > $x-$v-darwin-amd64.sha1
 	done
 done
 
-
-for i in {0..10}
+for v in $@
 do
-	wget  https://download.pingcap.org/tidb-v3.0.$i-linux-amd64.tar.gz
-	tar -xzf  tidb-v3.0.$i-linux-amd64.tar.gz
+	wget  https://download.pingcap.org/tidb-$v-linux-amd64.tar.gz
+	tar -xzf  tidb-$v-linux-amd64.tar.gz
 	for x in tidb tikv pd
 	do
-		tar -C tidb-v3.0.$i-linux-amd64 -czf  $x-v3.0.$i-linux-amd64.tar.gz bin/$x-server
-		sha1sum $x-v3.0.$i-linux-amd64.tar.gz | awk '{print $1}' > $x-v3.0.$i-linux-amd64.sha1
+		tar -C tidb-$v-linux-amd64 -czf  $x-$v-linux-amd64.tar.gz bin/$x-server
+		sha1sum $x-$v-linux-amd64.tar.gz | awk '{print $1}' > $x-$v-linux-amd64.sha1
 	done
-	rm -rf tidb-v3.0.$i-linux-amd64
+	rm -rf tidb-$v-linux-amd64
 done
